+++ pkg/app/appserver/proc.go
@@ -22,6 +22,7 @@ import (
 	"github.com/skycoin/skywire/pkg/app/appcommon"
 	"github.com/skycoin/skywire/pkg/app/appdisc"
 	"github.com/skycoin/skywire/pkg/app/appnet"
+	"github.com/skycoin/skywire/pkg/routing"
 	"github.com/skycoin/skywire/pkg/skyenv"
 )
 
@@ -67,6 +68,9 @@ type Proc struct {
 
 	errMx sync.RWMutex
 	err   string
+
+	portMx sync.RWMutex
+	port   routing.Port
 
 	cmdStderr io.ReadCloser
 
@@ -382,6 +386,21 @@ func (p *Proc) SetError(appErr string) {
 	p.err = appErr
 }
 
+// SetAppPort sets the proc's connection port
+func (p *Proc) SetAppPort(port routing.Port) {
+	p.portMx.Lock()
+	defer p.portMx.Unlock()
+	p.port = port
+}
+
+// GetAppPort gets the proc's connection port
+func (p *Proc) GetAppPort() routing.Port {
+	p.portMx.Lock()
+	defer p.portMx.Unlock()
+
+	return p.port
+}
+
 // Error gets proc's error.
 func (p *Proc) Error() string {
 	p.errMx.RLock()
@@ -433,10 +452,10 @@ func (p *Proc) ConnectionsSummary() []ConnectionSummary {
 			})
 			return true
 		}
-
 		summaries = append(summaries, ConnectionSummary{
-			IsAlive:            skywireConn.IsAlive(),
-			Latency:            skywireConn.Latency(),
+			IsAlive: skywireConn.IsAlive(),
+			// Latency in summary is expected to be in ms and not ns so we change the base to ms
+			Latency:            time.Duration(skywireConn.Latency().Milliseconds()),
 			UploadSpeed:        skywireConn.UploadSpeed(),
 			DownloadSpeed:      skywireConn.DownloadSpeed(),
 			BandwidthSent:      skywireConn.BandwidthSent(),