+++ pkg/visor/visor.go
@@ -96,6 +96,10 @@ type Visor struct {
 	connectedHypervisors map[cipher.PubKey]bool // remote hypervisors the visor is currently connected to
 	allowedPorts         []int
 	allowedMX            *sync.RWMutex
+
+	pingConns    map[cipher.PubKey]ping
+	pingConnMx   *sync.Mutex
+	pingPcktSize int
 }
 
 // todo: consider moving module closing to the module system
@@ -119,7 +123,7 @@ func (v *Visor) MasterLogger() *logging.MasterLogger {
 }
 
 // NewVisor constructs new Visor.
-func NewVisor(ctx context.Context, conf *visorconfig.V1, restartCtx *restart.Context, autoPeer bool, autoPeerIP string) (*Visor, bool) {
+func NewVisor(ctx context.Context, conf *visorconfig.V1, restartCtx *restart.Context, autoPeer bool, autoPeerIP string, dmsgServer string) (*Visor, bool) {
 
 	v := &Visor{
 		log:                  conf.MasterLogger().PackageLogger("visor"),
@@ -131,6 +135,8 @@ func NewVisor(ctx context.Context, conf *visorconfig.V1, restartCtx *restart.Con
 		dtmReady:             make(chan struct{}),
 		stunReady:            make(chan struct{}),
 		connectedHypervisors: make(map[cipher.PubKey]bool),
+		pingConns:            make(map[cipher.PubKey]ping),
+		pingConnMx:           new(sync.Mutex),
 	}
 	v.isServicesHealthy.init()
 
@@ -147,6 +153,9 @@ func NewVisor(ctx context.Context, conf *visorconfig.V1, restartCtx *restart.Con
 	ctx = context.WithValue(ctx, visorKey, v)
 	v.runtimeErrors = make(chan error)
 	ctx = context.WithValue(ctx, runtimeErrsKey, v.runtimeErrors)
+	if dmsgServer != "" {
+		ctx = context.WithValue(ctx, "dmsgServer", dmsgServer) //nolint
+	}
 	registerModules(v.MasterLogger())
 	var mainModule visorinit.Module
 	if v.conf.Hypervisor == nil {