+++ static/skywire-manager-src/src/app/components/vpn/pages/vpn-server-list/enter-vpn-server-password/enter-vpn-server-password.component.ts
@@ -1,6 +1,6 @@
 import { Component, ViewChild, ElementRef, OnInit, Inject } from '@angular/core';
 import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
-import { MatDialogRef, MatDialog, MatDialogConfig, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatLegacyDialogRef as MatDialogRef, MatLegacyDialog as MatDialog, MatLegacyDialogConfig as MatDialogConfig, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog';
 
 import { AppConfig } from 'src/app/app.config';