+++ static/skywire-manager-src/src/app/components/vpn/vpn-helpers.ts
@@ -1,4 +1,4 @@
-import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { MatLegacyDialog as MatDialog, MatLegacyDialogRef as MatDialogRef } from '@angular/material/legacy-dialog';
 import { Router } from '@angular/router';
 import { map, mergeMap } from 'rxjs/operators';
 import { Observable, of } from 'rxjs';